added BaseSocket parent for CSocket and SSocket, CSocket init
