+++ src/client/networking/CSocket.cpp
@@ -1,6 +1,1 @@
 #include "CSocket.h"
-
-CSocket::CSocket(intptr_t sockFd) {
-	this->sockFd = sockFd;
-	this->state = OPEN;
-}
